ENH: datetime-autounit: Detect unit from string arrays as well
Also move datetime string functions to their own source file.
