+++ numpy/core/SConscript
@@ -439,6 +439,7 @@ if ENABLE_SEPARATE_COMPILATION:
         pjoin('src', 'multiarray', 'hashdescr.c'),
         pjoin('src', 'multiarray', 'arrayobject.c'),
         pjoin('src', 'multiarray', 'datetime.c'),
+        pjoin('src', 'multiarray', 'datetime_strings.c'),
         pjoin('src', 'multiarray', 'datetime_busday.c'),
         pjoin('src', 'multiarray', 'datetime_busdaycal.c'),
         pjoin('src', 'multiarray', 'numpyos.c'),
