+++ numpy/core/src/multiarray/scalartypes.c.src
@@ -22,6 +22,7 @@
 #include "common.h"
 #include "scalartypes.h"
 #include "_datetime.h"
+#include "datetime_strings.h"
 
 NPY_NO_EXPORT PyBoolScalarObject _PyArrayScalar_BoolValues[] = {
     {PyObject_HEAD_INIT(&PyBoolArrType_Type) 0},
@@ -618,7 +619,7 @@ datetimetype_repr(PyObject *self)
         return NULL;
     }
 
-    if (make_iso_8601_date(&dts, iso, sizeof(iso), 1,
+    if (make_iso_8601_datetime(&dts, iso, sizeof(iso), 1,
                             scal->obmeta.base, -1) < 0) {
         return NULL;
     }
@@ -679,7 +680,7 @@ datetimetype_str(PyObject *self)
         return NULL;
     }
 
-    if (make_iso_8601_date(&dts, iso, sizeof(iso), 1,
+    if (make_iso_8601_datetime(&dts, iso, sizeof(iso), 1,
                             scal->obmeta.base, -1) < 0) {
         return NULL;
     }
